pray

+++ src/main/java/net/ccbluex/liquidbounce/ui/client/clickgui/newVer/element/module/value/impl/IntElement.kt
@@ -32,7 +32,7 @@ class IntElement(val savedValue: IntegerValue): ValueElement<Int>(savedValue) {
         Fonts.font40.drawString(savedValue.minimum.toString(), 
                                 x + width - 30F - sliderWidth - maxLength - minLength - valueDisplay, 
                                 y + 10F - Fonts.font40.FONT_HEIGHT / 2F, -1)
-        slider.setValue(savedValue.get(), savedValue.minimum, savedValue.maximum)
+        slider.setValue(savedValue.get().toFloat(), savedValue.minimum.toFloat(), savedValue.maximum.toFloat())
         slider.onDraw(x + width - 20F - sliderWidth - maxLength - valueDisplay, y + 10F, sliderWidth)
         RenderUtils.originalRoundedRect(x + width - 5F - valueDisplay, y + 2F, x + width - 10F, y + 18F, 4F, ColorManager.button.rgb)
         RenderUtils.customRounded(x + width - 18F, y + 2F, x + width - 10F, y + 18F, 0F, 4F, 4F, 0F, ColorManager.buttonOutline.rgb)
